+++ R/interval.R
@@ -43,9 +43,9 @@ interval_commit_metric <- function(project_git,dt_range,metric_function,
 #' commit interval on the form `commitsha1-commitsha2`
 #' @return a numeric vector of `metric_function` values for each commit interval specified in `dt_range`
 #' @export
-get_date_from_commit_hash <- function(git_log,commit_hash){
+get_date_from_commit_hash <- function(git_log,p_commit_hash){
   #data.commit <- NULL # due to NSE notes in R CMD check
-  return(git_log[data.commit == commit_hash]$data.AuthorDate[1])
+  return(git_log[commit_hash == p_commit_hash]$author_datetimetz[1])
 }
 
 #' @importFrom magrittr %>%